Update homepage to BS3
Instead of inhertiting from base.html, home.html has base.html code
included.  If home.html extends base.html, the sidebar space is there
because of "col-md-9" in base.html.  Changed "col-md-9" to "col-md-12"
to make the content take up the sidebar space as well.  There might be
a better fix for this...

Updated sticky-footer.html to give more space between the content and
the footer (BS3 messed this up, it seems)

Removed the responsive css file since it's implemented in BS3 natively